+++ CHANGES.md
@@ -68,6 +68,7 @@ Changes in 13.0
 - JS update Chart.js to version 4.5.1 in assets/js/Chart.js/
 - JS add Previous Next Student plugin to core in assets/js/jquery/previous-next-student/, warehouse.js & Gruntfile.js
 - Remove List of words ignored during search operations in ListOutput.fnc.php
+- JS deprecate LO_search() function in warehouse.js, csp/functions.js & ListOutput.fnc.php
 
 Changes in 12.9.3
 -----------------